feat(models): reconcile Mistral Nova xAI and Llama catalogs
