+++ src/core/providers/mistral/model_catalog.rs
@@ -1,0 +1,289 @@
+use std::collections::HashMap;
+
+use crate::core::types::model::ModelInfo;
+
+struct MistralModelSpec {
+    id: &'static str,
+    name: &'static str,
+    max_context_length: u32,
+    supports_streaming: bool,
+    supports_tools: bool,
+    supports_multimodal: bool,
+    input_cost_per_1k_tokens: f64,
+    output_cost_per_1k_tokens: f64,
+    alias_for: Option<&'static str>,
+}
+
+pub(super) fn mistral_model_catalog() -> Vec<ModelInfo> {
+    MODEL_SPECS.iter().map(model_info).collect()
+}
+
+fn model_info(spec: &MistralModelSpec) -> ModelInfo {
+    let metadata = spec
+        .alias_for
+        .map(|alias_for| HashMap::from([("alias_for".to_string(), serde_json::json!(alias_for))]))
+        .unwrap_or_default();
+
+    ModelInfo {
+        id: spec.id.to_string(),
+        name: spec.name.to_string(),
+        provider: "mistral".to_string(),
+        max_context_length: spec.max_context_length,
+        max_output_length: None,
+        supports_streaming: spec.supports_streaming,
+        supports_tools: spec.supports_tools,
+        supports_multimodal: spec.supports_multimodal,
+        input_cost_per_1k_tokens: Some(spec.input_cost_per_1k_tokens),
+        output_cost_per_1k_tokens: Some(spec.output_cost_per_1k_tokens),
+        currency: "USD".to_string(),
+        capabilities: vec![],
+        created_at: None,
+        updated_at: None,
+        metadata,
+    }
+}
+
+const MODEL_SPECS: &[MistralModelSpec] = &[
+    MistralModelSpec {
+        id: "mistral-large-2512",
+        name: "Mistral Large 3 2512",
+        max_context_length: 262144,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.0005,
+        output_cost_per_1k_tokens: 0.0015,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "mistral-medium-3-5",
+        name: "Mistral Medium 3.5",
+        max_context_length: 256000,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.0015,
+        output_cost_per_1k_tokens: 0.0075,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "mistral-medium-2508",
+        name: "Mistral Medium 3.1 2508",
+        max_context_length: 131072,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.0004,
+        output_cost_per_1k_tokens: 0.002,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "mistral-small-2603",
+        name: "Mistral Small 4 2603",
+        max_context_length: 256000,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.00015,
+        output_cost_per_1k_tokens: 0.0006,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "mistral-small-2506",
+        name: "Mistral Small 3.2 2506",
+        max_context_length: 256000,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.0001,
+        output_cost_per_1k_tokens: 0.0003,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "mistral-small-4",
+        name: "Mistral Small 4 (alias -> mistral-small-2603)",
+        max_context_length: 256000,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.00015,
+        output_cost_per_1k_tokens: 0.0006,
+        alias_for: Some("mistral-small-2603"),
+    },
+    MistralModelSpec {
+        id: "magistral-medium-2509",
+        name: "Magistral Medium 1.2 2509",
+        max_context_length: 40000,
+        supports_streaming: true,
+        supports_tools: false,
+        supports_multimodal: false,
+        input_cost_per_1k_tokens: 0.002,
+        output_cost_per_1k_tokens: 0.005,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "magistral-medium-1-2",
+        name: "Magistral Medium 1.2 (alias -> magistral-medium-2509)",
+        max_context_length: 40000,
+        supports_streaming: true,
+        supports_tools: false,
+        supports_multimodal: false,
+        input_cost_per_1k_tokens: 0.002,
+        output_cost_per_1k_tokens: 0.005,
+        alias_for: Some("magistral-medium-2509"),
+    },
+    MistralModelSpec {
+        id: "magistral-small-2509",
+        name: "Magistral Small 1.2 2509",
+        max_context_length: 40000,
+        supports_streaming: true,
+        supports_tools: false,
+        supports_multimodal: false,
+        input_cost_per_1k_tokens: 0.0005,
+        output_cost_per_1k_tokens: 0.0015,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "magistral-small-1-2",
+        name: "Magistral Small 1.2 (alias -> magistral-small-2509)",
+        max_context_length: 40000,
+        supports_streaming: true,
+        supports_tools: false,
+        supports_multimodal: false,
+        input_cost_per_1k_tokens: 0.0005,
+        output_cost_per_1k_tokens: 0.0015,
+        alias_for: Some("magistral-small-2509"),
+    },
+    MistralModelSpec {
+        id: "ministral-14b-2512",
+        name: "Ministral 14B 2512",
+        max_context_length: 262144,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.0001,
+        output_cost_per_1k_tokens: 0.0001,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "ministral-8b-2512",
+        name: "Ministral 8B 2512",
+        max_context_length: 262144,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.00005,
+        output_cost_per_1k_tokens: 0.00005,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "ministral-3b-2512",
+        name: "Ministral 3B 2512",
+        max_context_length: 131072,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: false,
+        input_cost_per_1k_tokens: 0.00004,
+        output_cost_per_1k_tokens: 0.00004,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "pixtral-large-2411",
+        name: "Pixtral Large 2411",
+        max_context_length: 131072,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.002,
+        output_cost_per_1k_tokens: 0.006,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "pixtral-12b-2409",
+        name: "Pixtral 12B 2409",
+        max_context_length: 128000,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.00015,
+        output_cost_per_1k_tokens: 0.00015,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "devstral-2512",
+        name: "Devstral 2512",
+        max_context_length: 262144,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: false,
+        input_cost_per_1k_tokens: 0.0004,
+        output_cost_per_1k_tokens: 0.002,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "devstral-2-2512",
+        name: "Devstral 2 2512 (alias -> devstral-2512)",
+        max_context_length: 262144,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: false,
+        input_cost_per_1k_tokens: 0.0004,
+        output_cost_per_1k_tokens: 0.002,
+        alias_for: Some("devstral-2512"),
+    },
+    MistralModelSpec {
+        id: "mistral-nemo-12b",
+        name: "Mistral Nemo 12B",
+        max_context_length: 131072,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: false,
+        input_cost_per_1k_tokens: 0.00015,
+        output_cost_per_1k_tokens: 0.00015,
+        alias_for: None,
+    },
+    MistralModelSpec {
+        id: "mistral-large",
+        name: "Mistral Large (alias -> mistral-large-2512)",
+        max_context_length: 262144,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.0005,
+        output_cost_per_1k_tokens: 0.0015,
+        alias_for: Some("mistral-large-2512"),
+    },
+    MistralModelSpec {
+        id: "mistral-small",
+        name: "Mistral Small (alias -> mistral-small-2603)",
+        max_context_length: 256000,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.00015,
+        output_cost_per_1k_tokens: 0.0006,
+        alias_for: Some("mistral-small-2603"),
+    },
+    MistralModelSpec {
+        id: "mistral-medium",
+        name: "Mistral Medium (alias -> mistral-medium-3-5)",
+        max_context_length: 256000,
+        supports_streaming: true,
+        supports_tools: true,
+        supports_multimodal: true,
+        input_cost_per_1k_tokens: 0.0015,
+        output_cost_per_1k_tokens: 0.0075,
+        alias_for: Some("mistral-medium-3-5"),
+    },
+    MistralModelSpec {
+        id: "mistral-embed",
+        name: "Mistral Embed",
+        max_context_length: 8192,
+        supports_streaming: false,
+        supports_tools: false,
+        supports_multimodal: false,
+        input_cost_per_1k_tokens: 0.0001,
+        output_cost_per_1k_tokens: 0.0,
+        alias_for: None,
+    },
+];
