+++ src/core/providers/ollama/streaming.rs
@@ -317,6 +317,7 @@ fn response_to_chunks(response: ChatResponse) -> Vec<ChatChunk> {
                 role: Some(MessageRole::Assistant),
                 content: None,
                 thinking: None,
+                audio: None,
                 tool_calls: None,
                 function_call: None,
             },
@@ -353,6 +354,7 @@ fn response_to_chunks(response: ChatResponse) -> Vec<ChatChunk> {
                             role: None,
                             content: Some(chunk_text),
                             thinking: None,
+                            audio: None,
                             tool_calls: None,
                             function_call: None,
                         },
@@ -377,6 +379,7 @@ fn response_to_chunks(response: ChatResponse) -> Vec<ChatChunk> {
                     role: None,
                     content: None,
                     thinking: None,
+                    audio: None,
                     tool_calls: None,
                     function_call: None,
                 },
